# swagger2.yaml
swagger: "2.0"
info:
  description: "Sprint 3 "
  version: "1.0.0"
  title: "Gestion de pedidos"
  termsOfService: "http://ejemplo/tyc/"
  contact:
    email: "luis.deseta@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost"
basePath: "/api"
tags:
- name: Pedidos
  description: "detalles para generar el ID de pedido"
- name: Items
  description: "detalles cargar los platos en la tabla Items"
paths: 
  /order:
    post:
      tags: 
      - Pedidos
      summary: Gernera Pedido nuevo
      description: Genera un ID de pedido con status "Nuevo" en la tabla pedidos. Si el usuario ya tiene un pedido activo, devuelve el ID de pedido.
      responses:
        200:
          description: ID de pedido
        400:
          description: Error
    put:
      tags: 
      - Pedidos
      summary: Actualiza datos del Pedido
      description: Si el role del usuario es "ADMIN" Actualiza status, total, resumen de platos (nombre corto) y medio de pago. Si el role es "USER" solo puede actualizar medio de pago.
      parameters:
      - in: "body"
        name: "body"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/Pedido"
      responses:
        200:
          description: OK
        405:
          description: Elemento Invalido
  /item/{idOrder}:
    post:
      tags: 
      - Items
      summary: Gernera la lista de platos
      description: Carga los platos en la tabla Items elegidos por el usuario, debe estar asociado a un ID de Pedido valido
      parameters:
      - in: header
        name: token
        required: true
        type: integer
      - in: path
        name: idOrder
        required: true
        type: string
      - in: "body"
        name: body
        description: Parametros con detalle del plato que solicita el usuario
        required: true
        schema:
          $ref: "#/definitions/Item"
      responses:
        200:
          description: ID de Item
        401:
          description: Verificar número de orden o id producto
  /item:
    put:
      tags: 
      - Items
      summary: Actualiza un plato - Solo Admin
      description: Actualiza los datos de un plato= Plato, Cantidad y precio.
      parameters:
      - in: header
        name: token
        required: true
        type: integer
      - in: "body"
        name: body
        description: Parametros con detalle del plato que solicita el usuario
        required: true
        schema:
          $ref: "#/definitions/Item"
      responses:
        200:
          description: Pedido Actualizado
        400:
          description: Error
    get:
      tags: 
      - Items
      summary: Devuelve datos de pedido (Pedido+Items)
      description: Usuario "Admin", consulta los datos de todos los pedidos. Usuario "User", consulta datos de pedidos generados ese ID de usuario (via token)
      parameters:
      - in: header
        name: token
        required: true
        type: integer
      - in: "body"
        name: status
        description: User puede elegir tipo de status a consultar. Default "nuevo" 
        required: false
        schema:
          $ref: "#/definitions/Item"
      responses:
        200:
          description: Detalle del pedido/os
        400:
          description: Error
 
definitions:
  Pedido:
    type: "object"
    properties:
      pedido_id:
        type: "integer"
        format: "int64"
      payment:
        type: "string"
      status:
        type: "string"
  Item:
    type: "object"
    properties:
      plato_id:
        type: "integer"
        format: "int64"
      quantity:
        type: "integer"
        format: "int64"
      price:
        type: "integer"
        format: "int64"      
            
            